+++ doc/changelog.rst
@@ -48,7 +48,8 @@ Front
 
 - Fix bug: Could no change quantities of unorderable lines in the basket
 - Use display units when rendering product quantities
-- Basket views: Add display unit support
+- Add new provide category called `product_context_extra`
+  which can be used to add extra data to the product context.
 - It's now possible to re-order old order from order history
 - It's now possible for addons to extend front main menu using the
   new `front_menu_extender` provide. See :doc:`provides.rst` for more information.